% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QDA.R
\name{bhatbound}
\alias{bhatbound}
\title{Finite-sample Bhattacharyya bound between two Gaussians \code{0} and \code{1}.}
\usage{
bhatbound(m0, S0, m0hat, S0hat, m1hat, S1hat, p0 = 0.5, p1 = 0.5, l = 0.5)
}
\arguments{
\item{m0}{true mean vector of Gaussian \code{0}}

\item{S0}{true covariance matrix of Gaussian \code{0}}

\item{m0hat}{estimation of \code{m0}}

\item{S0hat}{estimation of \code{S0}}

\item{m1hat}{estimation of \code{m1}}

\item{S1hat}{estimation of \code{S1}}

\item{p0}{prior of Gaussian \code{0}}

\item{p1}{prior of Gaussian \code{1}}

\item{l}{parameter of the Chernoff bound}
}
\value{
Chernoff distance between the two Gaussians
}
\description{
Finite-sample Bhattacharyya bound between two Gaussians \code{0} and \code{1}.
}
