% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Miscellaneous.R
\name{ccov}
\alias{ccov}
\title{Complex covariance matrix.}
\usage{
ccov(x)
}
\arguments{
\item{x}{matrix whose rows are the points; real or complex}
}
\value{
the sample covariance matrix of \code{x}
}
\description{
Complex covariance matrix.
}
