% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Miscellaneous.R
\name{covcheck}
\alias{covcheck}
\title{Checks if covariance and pseudo-covariance matrices are compatible.}
\usage{
covcheck(C, P)
}
\arguments{
\item{C}{covariance matrix of the distribution}

\item{P}{pseudo-covariance matrix of the distribution}
}
\value{
two columns of eigenvalues; must be non-negative to be compatible
}
\description{
Checks if covariance and pseudo-covariance matrices are compatible.
}
