% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Miscellaneous.R
\name{diagtrans}
\alias{diagtrans}
\title{Linear transformation into identity and diagonal covariances.}
\usage{
diagtrans(A, B)
}
\arguments{
\item{A}{Hermitian matrix to be turned into identity}

\item{B}{Hermitian matrix to be turned into diagonal}
}
\value{
matrix that must be applied to turn \code{A} into identity and \code{B} into diagonal
}
\description{
Linear transformation into identity and diagonal covariances.
}
