% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quadratic.R
\name{halfspace}
\alias{halfspace}
\title{Half-space classification with \code{CVXR}.}
\usage{
halfspace(x, y, p = 1, C = 10, a = 1, A = diag(ncol(x)))
}
\arguments{
\item{x}{matrix whose rows are the instances}

\item{y}{vector of labels}

\item{p}{p-norm of the regularisation vector; usually set to 1 or 2}

\item{C}{regularisation parameter; usually set from 2 to 10}

\item{a}{norm-constraint of the convex set}

\item{A}{is the matrix that skewes the constraint}
}
\value{
vector containing the constant term that points to class \code{1}
}
\description{
Half-space classification with \code{CVXR}.
}
