% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metric.R
\name{KNN}
\alias{KNN}
\title{Finds the k nearest neighbours of same labels (Weinberger and Saul (2009), Section 3.1).}
\usage{
KNN(x, y, k)
}
\arguments{
\item{x}{is the set whose rows are the instances sorted by labels}

\item{y}{is the set of labels in class 'factor'}

\item{k}{is how many nearest neighbours to consider}
}
\description{
Finds the k nearest neighbours of same labels (Weinberger and Saul (2009), Section 3.1).
}
