% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Miscellaneous.R
\name{matplot}
\alias{matplot}
\title{Convert a table to a form suitable for \code{matrix plots} in \code{pgfplots}.}
\usage{
matplot(x)
}
\arguments{
\item{x}{matrix whose rows are the points; its first two columns must be categorical and contain all possible pairs of values}
}
\value{
a table for use in \code{matrix plot} in \code{pgfplots}
}
\description{
Convert a table to a form suitable for \code{matrix plots} in \code{pgfplots}.
}
