% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metric.R
\name{Metric}
\alias{Metric}
\title{Out-of-sample error rate for metric under different compression methods.}
\usage{
Metric(
  x,
  y,
  k = ncol(x),
  s = sort(sample(1:nrow(x), 0.8 * nrow(x))),
  t = 10,
  W = c("G", "O", "B", "F", "C")
)
}
\arguments{
\item{x}{matrix whose rows are the instances}

\item{y}{vector of labels}

\item{k}{projection dimension}

\item{s}{vector of indices to be used as training set}

\item{t}{number of random comrpessions to average}

\item{W}{which type of compression to use}
}
\description{
Out-of-sample error rate for metric under different compression methods.
}
