% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Miscellaneous.R
\name{pcov}
\alias{pcov}
\title{Pseudo-covariance matrix.}
\usage{
pcov(x)
}
\arguments{
\item{x}{matrix whose rows are the points; real or complex}
}
\value{
the sample pseudo-covariance matrix of \code{x}
}
\description{
Pseudo-covariance matrix.
}
