% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quadratic.R
\name{perceptron}
\alias{perceptron}
\title{Half-space classification with \code{CVXR}.}
\usage{
perceptron(x, y, p = 1, C = 10)
}
\arguments{
\item{x}{matrix whose rows are the instances}

\item{y}{vector of labels}

\item{p}{p-norm of the regularisation vector; usually set to 1 or 2}

\item{C}{regularisation parameter; usually set from 2 to 10}
}
\value{
vector containing the constant term that points to class \code{1}
}
\description{
Half-space classification with \code{CVXR}.
}
