% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QDA.R
\name{qdaresult}
\alias{qdaresult}
\title{QDA error rate when sampling only from one Gaussian.}
\usage{
qdaresult(
  m0,
  S0,
  m1,
  S1,
  p0 = 0.5,
  p1 = 0.5,
  l = 0.5,
  n = 10,
  N = 100,
  t = 1000
)
}
\arguments{
\item{m0}{true mean vector of Gaussian \code{0}}

\item{S0}{true covariance matrix of Gaussian \code{0}}

\item{m1}{true mean vector of Gaussian \code{1}}

\item{S1}{true covariance matrix of Gaussian \code{1}}

\item{p0}{prior of Gaussian \code{0}}

\item{p1}{prior of Gaussian \code{1}}

\item{l}{parameter of the Chernoff bound}

\item{n}{training sample size to estimate mean vectors and covariance matrices}

\item{N}{test sample size from Gaussian \code{0}}

\item{t}{number of iterations}
}
\value{
Average classification error
}
\description{
QDA error rate when sampling only from one Gaussian.
}
