% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quadratic.R
\name{qproj}
\alias{qproj}
\title{Convex optimization for projection onto convex set using CVXR.}
\usage{
qproj(
  X,
  a,
  C = diag(nrow = nrow(X), ncol = ncol(X)),
  nor = c("nuc", 1, 2, "inf", "fro")
)
}
\arguments{
\item{X}{matrix to be projected onto convex set}

\item{a}{norm-constraint of the convex set}

\item{C}{inverse matrix that skews the convex set}

\item{nor}{which norm to constrain}
}
\value{
the matrix in the convex set closest in norm \code{nor} to \code{X}
}
\description{
Convex optimization for projection onto convex set using CVXR.
}
