% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quadratic.R
\name{quadappend}
\alias{quadappend}
\title{Append a set with its quadratic terms}
\usage{
quadappend(x)
}
\arguments{
\item{x}{matrix whose rows are the instances}
}
\value{
matrix containing \code{x} with its quadratic terms
}
\description{
Append a set with its quadratic terms
}
