% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quadratic.R
\name{quadsgd}
\alias{quadsgd}
\title{Train quadratic classifier under projection with SGD}
\usage{
quadsgd(
  x,
  y,
  B = diag(ncol(x)),
  g = 0.5,
  epoch = 2,
  alpha = 1,
  A = diag(0, nrow(B)),
  b = rep(0, nrow(B)),
  c = 0,
  a = 3,
  nor = c("nuc", 1, 2, "inf", "fro")
)
}
\arguments{
\item{x}{matrix whose rows are the instances}

\item{y}{vector of labels}

\item{B}{projection matrix; defaults to no projection}

\item{g}{parameter of the smoothed hinge loss}

\item{epoch}{number of epochs in the SGD}

\item{alpha}{learning rate of the SGD; can be modified to vary for each iteration}

\item{A}{matrix with which to initialise SGD}

\item{b}{vector with which to initialise SGD}

\item{c}{constant with which to initialise SGD}

\item{a}{norm-constraint for the matrix in the quadratic boundary}

\item{nor}{which norm to constrain}
}
\description{
Train quadratic classifier under projection with SGD
}
