% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quadratic.R
\name{qucl}
\alias{qucl}
\title{Value of quadratic classifier for all instances.}
\usage{
qucl(x, A = diag(ncol(x)), b = rep(0, ncol(x)), c = 0)
}
\arguments{
\item{x}{matrix whose rows are the instances}

\item{A}{matrix in the quadratic form}

\item{b}{vector in the quadratic form}

\item{c}{constant in the quadratic form}
}
\value{
vector of values
}
\description{
Value of quadratic classifier for all instances.
}
\examples{
# direct label prediction

sign(qucl(matrix(rnorm(500),5)))
}
