% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Miscellaneous.R
\name{rcmnorm}
\alias{rcmnorm}
\title{Sample from multi-variate complex Gaussian distribution.}
\usage{
rcmnorm(n, m, C, P)
}
\arguments{
\item{n}{number of vectors to sample}

\item{m}{mean vector of the distribution}

\item{C}{covariance matrix of the distribution}

\item{P}{pseudo-covariance matrix of the distribution}
}
\value{
matrix whose rows are the sampled vectors
}
\description{
Sample from multi-variate complex Gaussian distribution.
}
