% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QDA.R
\name{reff}
\alias{reff}
\title{sample uniformly \code{d} eigenvalues given largest eigenvalue and effective rank}
\usage{
reff(d, l, r)
}
\arguments{
\item{d}{The number of dimensions.}

\item{l}{The largest eigenvalue.}

\item{r}{The desired effective rank.}
}
\value{
Vector of eigenvalues with desired effective rank.
}
\description{
sample uniformly \code{d} eigenvalues given largest eigenvalue and effective rank
}
