% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quadratic.R
\name{smhinge}
\alias{smhinge}
\title{Smoothed hinge loss}
\usage{
smhinge(x, g = 0.5)
}
\arguments{
\item{x}{input vector}

\item{g}{parameter}
}
\value{
vector of values
}
\description{
Smoothed hinge loss
}
