% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quadratic.R
\name{smhingeder}
\alias{smhingeder}
\title{Derivative of smoothed hinge loss}
\usage{
smhingeder(x, g = 0.5)
}
\arguments{
\item{x}{vector}

\item{g}{parameter}
}
\value{
vector of values
}
\description{
Derivative of smoothed hinge loss
}
