% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quadratic.R
\name{vec2mat}
\alias{vec2mat}
\title{Reconstruct matrix from vector.}
\usage{
vec2mat(x)
}
\arguments{
\item{x}{is the vector containing the cross terms}
}
\value{
a matrix
}
\description{
Reconstruct matrix from vector.
}
