% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Miscellaneous.R
\name{vecperm}
\alias{vecperm}
\title{Permutation of vectorization.}
\usage{
vecperm(m, n)
}
\arguments{
\item{m}{number of rows}

\item{n}{number of columns}
}
\value{
permutation matrix \code{P} such that \code{P\%*\%c(A)=c(t(A))}
}
\description{
Permutation of vectorization.
}
