% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quadratic.R
\name{vproj}
\alias{vproj}
\title{CVXR for projection of vector onto convex set.}
\usage{
vproj(X, a = 1, nor = Inf, C = rep(1, length(X)))
}
\arguments{
\item{X}{vector to be projected onto convex set}

\item{a}{norm-constraint of the convex set}

\item{nor}{which p-norm to constrain; defaults to infinity}

\item{C}{inverse of the vector that skews the convex set; defaults to no skew}
}
\value{
the vector in the convex set closest in norm \code{nor} to \code{X}
}
\description{
CVXR for projection of vector onto convex set.
}
