% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Miscellaneous.R
\name{WHmat}
\alias{WHmat}
\title{Create a \code{d*d} Walsh-Hadamard matrix.}
\usage{
WHmat(d)
}
\arguments{
\item{d}{number of dimensions}
}
\value{
a \code{d*d} matrix with the Walsh-Hadamard property
}
\description{
Create a \code{d*d} Walsh-Hadamard matrix.
}
\examples{
# 4*4 Walsh-Hadamard matrix

WHmat(4)

}
