% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quadratic.R
\name{qproj}
\alias{qproj}
\title{Projection of matrix onto convex set with \code{CVXR}.}
\usage{
qproj(X, a = 1, nor = c("nuc", 1, 2, "inf", "fro"), C = diag(ncol(X)))
}
\arguments{
\item{X}{matrix to be projected onto convex set}

\item{a}{norm-constraint of the convex set}

\item{nor}{which matrix-norm to constrain; defaults to nuclear}

\item{C}{inverse of the matrix that skews the convex set; defaults to no skew}
}
\value{
the matrix in the convex set closest in norm \code{nor} to \code{X}
}
\description{
Projection of matrix onto convex set with \code{CVXR}.
}
