% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quadratic.R
\name{quadsgd}
\alias{quadsgd}
\title{Train quadratic classifier under projection with SGD}
\usage{
quadsgd(
  x,
  y,
  B,
  g = 0.5,
  epoch = 2,
  alpha = 0.1,
  A = diag(0, nrow(B)),
  b = rep(0, nrow(B)),
  c = 0,
  a = 3,
  nor = c("nuc", 1, 2, "inf", "fro"),
  skew = T
)
}
\arguments{
\item{x}{matrix whose rows are the instances}

\item{y}{vector of labels}

\item{B}{projection matrix; defaults to no projection}

\item{g}{parameter of the smoothed hinge loss}

\item{epoch}{number of epochs in the SGD}

\item{alpha}{learning rate of the SGD; can be modified to vary for each iteration}

\item{A}{matrix with which to initialise SGD}

\item{b}{vector with which to initialise SGD}

\item{c}{constant with which to initialise SGD}

\item{a}{norm-constraint of the convex set}

\item{nor}{which norm to constrain; defaults to nuclear}

\item{skew}{whether to skew the compressed class or not}
}
\value{
list containing the following elements

\strong{\code{A}} matrix in the trained quadratic classifier

\strong{\code{b}} vector in the trained quadratic classifier

\strong{\code{c}} constant in the trained quadratic classifier

\strong{\code{error}} sequence of empirical errors during the SGD
}
\description{
Train quadratic classifier under projection with SGD
}
