% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QDA.R
\name{reff}
\alias{reff}
\title{Sample uniformly \code{d} eigenvalues given largest eigenvalue and effective rank.}
\usage{
reff(d, l, r)
}
\arguments{
\item{d}{number of dimensions}

\item{l}{largest eigenvalue}

\item{r}{desired effective rank}
}
\value{
vector of eigenvalues with desired effective rank
}
\description{
Sample uniformly \code{d} eigenvalues given largest eigenvalue and effective rank.
}
\examples{
#generates all ones

reff(5,1,5) 

#generates randomly

reff(5,1,3)

#generates all-but-one zeros

reff(5,1,1)
}
